+++ client/src/PDFFoldingRangeProvider.ts
@@ -5,91 +5,119 @@ export class PDFFoldingRangeProvider implements FoldingRangeProvider {
 		const ranges: FoldingRange[] = [];
 
 		let startObjLine = -1;
-		let startBracketLine = -1;
 		let startStreamLine = -1;
-		let startPdfNameLine = -1;
-		let startQLine = -1;
-		let startBTLine = -1;
-		let startBXLine = -1;
-		let startBDCLine = -1;
-		let startBMCLine = -1;
+		let startxrefLine = -1;
+		let startDictLines = [];
+		let startQLines = [];
+		let startBTLines = [];
+		let startBXLines = [];
+		let startMarkedContentLines = [];
 
 		for (let i = 0; i < document.lineCount; i++) {
 			const line = document.lineAt(i).text.trim();
 
-			if (line.startsWith('obj')) {
+			/// @TODO Remove any comments to leave just content. Handle strings
+			/// Especially important if a comment or string contains any of the markers!
+
+			// cannot be nested 
+			if (line.match('\\d+ \\d+ obj')) {
 				startObjLine = i;
+				startDictLines = []; // reset dictionaries
+				continue;
 			} else if (line.startsWith('endobj') && startObjLine >= 0) {
 				const r = new FoldingRange(startObjLine, i);
 				ranges.push(r);
 				startObjLine = -1;
+				continue;
 			}
 
-			if (line.startsWith('<<')) {
-				startBracketLine = i;
-			} else if (line.startsWith('>>') && startBracketLine >= 0) {
-				const r = new FoldingRange(startBracketLine, i);
+			// cannot be nested 
+			if (line.startsWith('xref')) {
+				startxrefLine = i;
+				startDictLines = []; // reset dictionaries
+				continue;
+			} else if (line.startsWith('trailer') && startxrefLine >= 0) {
+				const r = new FoldingRange(startxrefLine, i);
 				ranges.push(r);
-				startBracketLine = -1;
+				startxrefLine = -1;
+				continue;
 			}
 
+			// cannot be nested 
 			if (line.startsWith('stream')) {
 				startStreamLine = i;
-			} else if (
-				(line.startsWith('endstream') || i === document.lineCount - 1) &&
-				startStreamLine >= 0
-			) {
+				startDictLines = []; // reset dictionaries
+				startQLines = []; // reset q/Q operand pairs
+				startBTLines = []; // reset BT/ET operand pairs
+				startBXLines = []; // reset BX/EX operand pairs
+				startMarkedContentLines = []; // reset BDC/BMC/EMC operand pairs
+				continue;
+			} else if (line.startsWith('endstream') && (startStreamLine >= 0)) {
 				const r = new FoldingRange(startStreamLine, i);
 				ranges.push(r);
 				startStreamLine = -1;
+				startDictLines = []; // reset dictionaries
+				startQLines = []; // reset q/Q operand pairs
+				startBTLines = []; // reset BT/ET operand pairs
+				startBXLines = []; // reset BX/EX operand pairs
+				startMarkedContentLines = []; // reset BDC/BMC/EMC operand pairs
+				continue;
 			}
 
-			if (line.startsWith('/') && line.includes('<<')) {
-				startPdfNameLine = i;
-			} else if (line.startsWith('>>') && startPdfNameLine >= 0) {
-				const r = new FoldingRange(startPdfNameLine, i);
+			// Nestable. Avoid when an entire dict (both "<<" and ">>") are on 1 line
+			if (line.startsWith('<<') && !line.includes('>>')) {
+				startDictLines.push(i);
+				continue;
+			}
+			else if (line.startsWith('/') && line.includes('<<')  && !line.includes('>>')) {
+				startDictLines.push(i);
+				continue;
+			} else if (line.startsWith('>>') && (startDictLines.length > 0)) {
+				const r = new FoldingRange(startDictLines.pop(), i);
 				ranges.push(r);
-				startPdfNameLine = -1;
+				continue;
 			}
 
+			// No operands or other operators. Nestable.
 			if (line.startsWith('q')) {
-				startQLine = i;
-			} else if (line.startsWith('Q') && startQLine >= 0) {
-				const r = new FoldingRange(startQLine, i);
+				startQLines.push(i);
+				continue;
+			} else if (line.startsWith('Q') && (startQLines.length > 0)) {
+				const r = new FoldingRange(startQLines.pop(), i);
 				ranges.push(r);
-				startQLine = -1;
+				continue;
 			}
 
+			// No operands or other operators. Nestable.
 			if (line.startsWith('BT')) {
-				startBTLine = i;
-			} else if (line.startsWith('ET') && startBTLine >= 0) {
-				const r = new FoldingRange(startBTLine, i);
+				startBTLines.push(i);
+				continue;
+			} else if (line.startsWith('ET') && startBTLines.length > 0) {
+				const r = new FoldingRange(startBTLines.pop(), i);
 				ranges.push(r);
-				startBTLine = -1;
+				continue;
 			}
 
+			// No operands or other operators. Nestable.
 			if (line.startsWith('BX')) {
-				startBXLine = i;
-			} else if (line.startsWith('EX') && startBXLine >= 0) {
-				const r = new FoldingRange(startBXLine, i);
+				startBXLines.push(i);
+				continue;
+			} else if (line.startsWith('EX') && startBXLines.length > 0) {
+				const r = new FoldingRange(startBXLines.pop(), i);
 				ranges.push(r);
-				startBXLine = -1;
+				continue;
 			}
 
-			if (line.startsWith('BDC')) {
-				startBDCLine = i;
-			} else if (line.startsWith('BMC')) {
-				startBMCLine = i;
-			} else if (line.startsWith('EMC')) {
-				if (startBDCLine >= 0) {
-					const r = new FoldingRange(startBDCLine, i);
-					ranges.push(r);
-					startBDCLine = -1;
-				} else if (startBMCLine >= 0) {
-					const r = new FoldingRange(startBMCLine, i);
-					ranges.push(r);
-					startBMCLine = -1;
-				}
+			// Nestable. Supports operands on same line as operator: 
+			//   tag properties BDC
+			//   tag BMC
+			if ((line.includes('BDC')) || (line.includes('BMC'))) {
+				startMarkedContentLines.push(i);
+				continue;
+			} else if (line.startsWith('EMC') && (startMarkedContentLines.length > 0)) {
+				const r = new FoldingRange(startMarkedContentLines.pop(), i);
+				ranges.push(r);
+				continue;
 			}
 		}
 		return ranges;