+++ Changelog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.1.3 - 2023-08-xx
+- (FIXED/ADDED) bracket matching
+- (ADDED) Additional validation checks of conventional cross reference tables
 - 
 
 ## 0.1.2 - 2023-08-15