+++ Changelog.md
@@ -6,12 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.1.3 - 2023-08-xx
-- (FIXED/ADDED) bracket matching
-- (ADDED) Additional validation checks of conventional cross reference tables
-- (ADDED) Auto-complete and auto-closing for arrays (`[`,`]`), literal strings (`(`/`)`), hex strings (`<`/`>`) and PostScript brackets (`{`/`}`). Dictionaries (`<<`,`>>`)
+- (IMPROVED) bracket matching for dictionaries (`<<`,`>>`), arrays (`[`,`]`), and PostScript brackets (`{`/`}`). 
+- (IMPROVED) auto-indent and auto-outdent for dictionaries (`<<`,`>>`), arrays (`[`,`]`), hex strings (`<`/`>`) and PostScript brackets (`{`/`}`). 
+- (ADDED) auto-complete and auto-closing for dictionaries (`<<`,`>>`), arrays (`[`,`]`), literal strings (`(`/`)`), hex strings (`<`/`>`) and PostScript brackets (`{`/`}`). 
 - (ADDED) LSP semantic token processor used by "go to" functionality to ensure correct token is located
 - (IMPROVED) TextMate grammar updates for syntax highlighting for PDFs with binary data
 - (IMPROVED) Folding support for objects, streams, conventional cross-reference tables and paired content stream operators 
+- (ADDED and FIXED) Additional validation checks of conventional cross reference tables
 
 
 ## 0.1.2 - 2023-08-15